requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.10
disnake==2.3.0
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
python-dotenv==0.19.2
yarl==1.7.2
